refactor(web): the comment compose bubble leaves SpatialEditor's return as CommentComposeOverlay
Step 2 of the JSX split: the draft bubble's placement (commentDraftBox
over canvas-render's own placer), chrome (commentComposeStyle — exported,
CommentThreadCard wears the same chrome), and the commit's three arms
(rewrite text / open a thread when a passage or node-set anchor rides
along / create a flat comment; blank commit = cancel, ADR-0025) move to
comment-compose-overlay.tsx. The caller keeps the obstacle closure and
passes its value. Behavior-preserving; comment
create/edit/reply/placement browser suites 19/19.
